Add comprehensive tests for core functionality and integration
- Introduced unit tests for OutputState, OutputWriter, ContentResolver, Session, Store, UniqueStore, and utility functions.
- Implemented integration tests for basic, Next.js, and Vite examples to ensure correct output generation and asset handling.
- Enhanced test coverage for session management, asset storage, and schema validation.
- Verified that builds do not reuse unique values or emitted assets from previous configurations.
- Added tests for file watching functionality to ensure rebuilds occur on content or config changes.
