+++ src/core/output.ts
@@ -3,12 +3,57 @@ import { join, relative } from 'node:path'
 
 import { logger as defaultLogger } from './logger'
 
-import type { Collections, Output } from '../types'
+import type { Collections } from '../collections'
 import type { AssetStore } from './assets'
 import type { Logger } from './logger'
 import type { OutputState } from './output-state'
 
 const isProduction = process.env.NODE_ENV === 'production'
+
+/**
+ * Output options
+ */
+export interface Output {
+  /**
+   * The output directory of the data files (relative to config file).
+   * @default '.velite'
+   */
+  data: string
+  /**
+   * The directory of the assets (relative to config file),
+   * should be served statically by the app
+   * `--clean` will automatically clear this directory
+   * @default 'public/static'
+   */
+  assets: string
+  /**
+   * The public base path of the assets
+   * @default '/static/'
+   * @example
+   * '/' -> '/image.png'
+   * '/static/' -> '/static/image.png'
+   * './static/' -> './static/image.png'
+   * 'https://cdn.example.com/' -> 'https://cdn.example.com/image.png'
+   */
+  base: '/' | `/${string}/` | `.${string}/` | `${string}:${string}/`
+  /**
+   * This option determines the name of each output asset.
+   * The asset will be written to the directory specified in the `output.assets` option.
+   * You can use `[name]`, `[hash]` and `[ext]` template strings with specify length.
+   * @default '[name]-[hash:8].[ext]'
+   */
+  name: string
+  /**
+   * Whether to clean the output directories before build
+   * @default false
+   */
+  clean: boolean
+  /**
+   * Output entry file format
+   * @default 'esm'
+   */
+  format: 'esm' | 'cjs'
+}
 
 export interface OutputWriterDeps {
   writeFile(path: string, content: string): Promise<void>
