+++ src/core/session.ts
@@ -4,7 +4,8 @@ import { createLogger, logger as defaultLogger } from './logger'
 import { createOutputState } from './output-state'
 import { createSessionStore } from './store'
 
-import type { Config, Loader } from '../types'
+import type { Config } from '../config'
+import type { Loader } from '../loaders/types'
 import type { FileCache } from './file-cache'
 import type { Logger } from './logger'
 import type { OutputState } from './output-state'
