+++ src/loaders/json.ts
@@ -1,4 +1,4 @@
-import { defineLoader } from '../types'
+import { defineLoader } from './types'
 
 export default defineLoader({
   // name: 'json',
