+++ src/schemas/index.ts
@@ -34,3 +34,8 @@ export { z }
 export type Schema<Output = unknown, Input = unknown> = z.ZodType<Output, Input>
 export type ZodType<Output = unknown, Input = unknown> = z.ZodType<Output, Input>
 export type infer<T extends z.ZodType> = z.infer<T>
+
+/**
+ * Define a schema (identity function for type inference)
+ */
+export const defineSchema = <T extends () => Schema>(fn: T): T => fn
