+++ test/core/output-state.tests.ts
@@ -1,7 +1,7 @@
 import { ok, strictEqual } from 'node:assert'
 import { describe, it } from 'node:test'
 
-import { createOutputState } from '../src/core/output-state'
+import { createOutputState } from '../../src/core/output-state'
 
 describe('OutputState', () => {
   it('starts empty', () => {
