+++ test/core/watch.tests.ts
